correct the error "KeyError: np.int16(-28989)" that shows up when generating for PLZ 94315 (bc it is big I suppose)

+++ src/pylovo/database/clustering_mixin.py
@@ -410,7 +410,7 @@ class ClusteringMixin(BaseMixin, ABC):
                                 %(cl)s, %(tl)s,
                                 false);"""
         cost_df = pd.read_sql_query(cost_query, con=self.conn, params={"cl": consumer_list, "tl": transformer_list},
-                                    dtype={"start_vid": np.int16, "end_vid": np.int16, "agg_cost": np.int16}, )
+                                     dtype={"start_vid": np.int32, "end_vid": np.int32, "agg_cost": np.float64}, )
 
         return cost_df
